Create Pallindrome
A palindrome number - number that remains the same after reversing each digit of that number. A prime number - number that is divisible by only one or itself. A number that satisfies both the properties is said to be PalPrime Number. 

Create a class PalPrime with a parameterised constructor PalPrime(int number, String message).

Given an positive integer array of numbers, you have to traverse the array and print the message “Number ___ is Prime/Palindrome/PalPrime”. 

Note: Message should be printed via constructor of PalPrime class. 

Input :
Array: [1, 34543, 565, 727, 10099] 

Output -> Predict the output